requirements.txt
click
requests
gitignorefile
coloredlogs
tqdm
astor
joblib
pyarrow
pandas
psutil
requirements-parser
setuptools
packaging
pytest
python-dotenv
humanfriendly